+++ Nyrkky/Source/BunkerBound/Engine/Render/GenericObjectRenderData.h
@@ -11,15 +11,15 @@ public:
 	void Init(int PosX, int PosY) override;
 	bool Activate(int PosX, int PosY) override;
 
-	void SetTexture(size_t texID)
+	void SetTexture(uint32_t texID)
 	{
 		_texID = texID;
 	}
 
 protected:
 	void AddVertex(std::vector<float>& x, float v1, float v2, float v3, float v4, float texIndex) override;
-	void AddIndex(std::vector<unsigned int>& x, float v1, float v2, float v3) override;
+	void AddIndex(std::vector<unsigned int>& x, int v1, int v2, int v3) override;
 
-	size_t _texID;
+	uint32_t _texID;
 };
